+++ contracts/mocks/MockAssetHandler.sol
@@ -20,7 +20,7 @@ contract MockAssetHandler is StorageLayoutV1 {
         CashGroup.setCashGroupStorage(id, cg);
     }
 
-    function buildCashGroupView(uint256 currencyId)
+    function buildCashGroupView(uint16 currencyId)
         public
         view
         returns (CashGroupParameters memory)
