+++ contracts/mocks/MockMarket.sol
@@ -25,7 +25,7 @@ contract MockMarket is StorageLayoutV1 {
         CashGroup.setCashGroupStorage(id, cg);
     }
 
-    function buildCashGroupView(uint256 currencyId)
+    function buildCashGroupView(uint16 currencyId)
         public
         view
         returns (CashGroupParameters memory)
